requirements.txt
Django>=5.0,<6.0
Pillow>=10.0.0
django-crispy-forms>=2.1
crispy-bootstrap5>=2024.1
django-allauth>=0.61.0
requests>=2.0
PyJWT>=2.0
cryptography>=42.0
